Add back support for stage.remap_qualname()
